add d8 status V4
