+++ internal/status/cmd/status.go
@@ -1,3 +1,19 @@
+/*
+Copyright 2025 Flant JSC
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
 package status
 
 import (
@@ -11,18 +27,18 @@ import (
 	"k8s.io/client-go/rest"
 	"k8s.io/kubectl/pkg/util/templates"
 
-	"github.com/deckhouse/deckhouse-cli/internal/status/statusresult"
+	"github.com/deckhouse/deckhouse-cli/internal/status/tools/statusresult"
 	"github.com/deckhouse/deckhouse-cli/internal/utilk8s"
 
-	"github.com/deckhouse/deckhouse-cli/internal/status/commands/clusteralerts"
-	"github.com/deckhouse/deckhouse-cli/internal/status/commands/cni_modules"
-	"github.com/deckhouse/deckhouse-cli/internal/status/commands/edition"
-	"github.com/deckhouse/deckhouse-cli/internal/status/commands/masters"
-	"github.com/deckhouse/deckhouse-cli/internal/status/commands/pods"
-	"github.com/deckhouse/deckhouse-cli/internal/status/commands/queue"
-	"github.com/deckhouse/deckhouse-cli/internal/status/commands/registry"
-	"github.com/deckhouse/deckhouse-cli/internal/status/commands/releases"
-	"github.com/deckhouse/deckhouse-cli/internal/status/commands/settings"
+	"github.com/deckhouse/deckhouse-cli/internal/status/objects/clusteralerts"
+	"github.com/deckhouse/deckhouse-cli/internal/status/objects/cni_modules"
+	"github.com/deckhouse/deckhouse-cli/internal/status/objects/edition"
+	"github.com/deckhouse/deckhouse-cli/internal/status/objects/masters"
+	"github.com/deckhouse/deckhouse-cli/internal/status/objects/pods"
+	"github.com/deckhouse/deckhouse-cli/internal/status/objects/queue"
+	"github.com/deckhouse/deckhouse-cli/internal/status/objects/registry"
+	"github.com/deckhouse/deckhouse-cli/internal/status/objects/releases"
+	"github.com/deckhouse/deckhouse-cli/internal/status/objects/settings"
 )
 
 var statusLong = templates.LongDesc(`
