--- internal/status/statusresult/statusresult.go
@@ -1,8 +1,0 @@
-package statusresult
-
-type StatusResult struct {
-	Title  string
-	Output string
-	Level  int
-	IsWarn bool
-}
